Test with perspective transform

+++ elikos_localization/src/localization/algorithm/DBSCAN.h
@@ -12,20 +12,20 @@
 
 namespace DBSCAN
 {
-void DBSCAN(const std::vector<cv::Point2d> &dataset, double epsilon, int minPts, std::vector<int>& clusterMemberships);
+void DBSCAN(const std::vector<cv::Point2f> &dataset, double epsilon, int minPts, std::vector<int>& clusterMemberships);
 
-Eigen::MatrixXd calculateDistMap (const std::vector<cv::Point2d> &dataset);
+Eigen::MatrixXd calculateDistMap (const std::vector<cv::Point2f> &dataset);
 
-Eigen::MatrixXd calculateDistMap (const std::vector<cv::Point2d> &dataset1, const std::vector<cv::Point2d> &dataset2);
+Eigen::MatrixXd calculateDistMap (const std::vector<cv::Point2f> &dataset1, const std::vector<cv::Point2f> &dataset2);
 
-double calculateDistance (const std::vector<cv::Point2d> &dataset, int i, int j);
+double calculateDistance (const std::vector<cv::Point2f> &dataset, int i, int j);
 
-double calculateDistance (const std::vector<cv::Point2d> &dataset1, const std::vector<cv::Point2d> &dataset2, int i, int j);
+double calculateDistance (const std::vector<cv::Point2f> &dataset1, const std::vector<cv::Point2f> &dataset2, int i, int j);
 
 std::vector<int> regionQuery (Eigen::MatrixXd distMap, double epsilon, int i);
 
 void parseClusterMemberShips(const std::vector<int>& clusterMemberShips,
-                                                 const std::vector<cv::Point2d>& intersections,
+                                                 const std::vector<cv::Point2f>& intersections,
                                                  std::vector<localization::Intersection>& parsedClusters);
 
 }
