+++ elikos_localization/src/localization/context/Line.cpp
@@ -10,23 +10,23 @@ namespace localization
 Line::Line(float rho, float theta)
     : rho_(std::abs(rho))
 {
-    orientation_ = cv::Vec2d(sinf(theta), cosf(theta));
-    centroid_ = cv::Point2d(rho * orientation_[1], rho * orientation_[0]);
+    orientation_ = cv::Vec2f(sinf(theta), cosf(theta));
+    centroid_ = cv::Point2f(rho * orientation_[1], rho * orientation_[0]);
 }
 
-Line::Line(float rho, const cv::Vec2d& orientation)
+Line::Line(float rho, const cv::Vec2f& orientation)
     : rho_(std::abs(rho)), orientation_(orientation)
 {
-    centroid_ = cv::Point2d(rho * orientation_[1], rho * orientation_[0]);
+    centroid_ = cv::Point2f(rho * orientation_[1], rho * orientation_[0]);
 }
 
-void Line::rotate(double rotation)
+void Line::rotate(float rotation)
 {
-    double x = orientation_[0];
-    double y = orientation_[1];
+    float x = orientation_[0];
+    float y = orientation_[1];
 
-    orientation_[0] = x * cos(rotation) - y * sin(rotation);
-    orientation_[1] = x * sin(rotation) + y * cos(rotation);
+    orientation_[0] = x * cosf(rotation) - y * sinf(rotation);
+    orientation_[1] = x * sinf(rotation) + y * cosf(rotation);
 }
 
 void Line::inverseOrientation()
@@ -35,7 +35,7 @@ void Line::inverseOrientation()
 }
 
 
-bool Line::findIntersection(const Line& otherLine, cv::Point2d& intersection) const
+bool Line::findIntersection(const Line& otherLine, cv::Point2f& intersection) const
 {
     const int x = 0;
     const int y = 1;
@@ -44,16 +44,16 @@ bool Line::findIntersection(const Line& otherLine, cv::Point2d& intersection) co
     //cv::Point2f A = { 100.0, 100.0 };
     //cv::Point2f B = { 200.0, 200.0};
 
-    cv::Point2d A = centroid_;
-    cv::Point2d B = otherLine.centroid_;
+    cv::Point2f A = centroid_;
+    cv::Point2f B = otherLine.centroid_;
 
-    cv::Vec2d dc = B - A;
+    cv::Vec2f dc = B - A;
 
     //cv::Vec2f u = { 2.0, 1.0 };
     //cv::Vec2f v = { -1.0, -2.0 };
 
-    cv::Vec2d u = orientation_;
-    cv::Vec2d v = otherLine.orientation_;
+    cv::Vec2f u = orientation_;
+    cv::Vec2f v = otherLine.orientation_;
 
     u = cv::normalize(u);
     v = cv::normalize(v);
@@ -67,7 +67,7 @@ bool Line::findIntersection(const Line& otherLine, cv::Point2d& intersection) co
     if (det != 0.0) {
 
         double d = (dc[x] * v[y] - dc[y] * v[x]) / det;
-        intersection = A + (cv::Point2d)(d * u);
+        intersection = A + (cv::Point2f)(d * u);
 
     } else {
         return false;
