+++ elikos_localization/src/localization/context/LineGroup.h
@@ -21,12 +21,12 @@ public:
     Line convertToLine() const;
 
     inline const std::vector<Line*>& getLines() const;
-    inline const cv::Vec2d& getAvgOrientation() const;
+    inline const cv::Vec2f& getAvgOrientation() const;
     inline double getAvgRho() const;
 
 private :
-    cv::Vec2d avgOrientation_;
-    double avgRho_;
+    cv::Vec2f avgOrientation_;
+    float avgRho_;
 
     std::vector<Line*> lines_;
 
@@ -38,7 +38,7 @@ inline const std::vector<Line*>& LineGroup::getLines() const
    return lines_;
 }
 
-inline const cv::Vec2d& LineGroup::getAvgOrientation() const
+inline const cv::Vec2f& LineGroup::getAvgOrientation() const
 {
    return avgOrientation_;
 }
